MathGenX-Ai/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
pydantic
requests
python-dotenv 
google-genai
anthropic
# flask